final before unit testing

+++ Model/RecipeModel.swift
@@ -22,13 +22,13 @@ struct Hit: Codable {
 
 struct Recipe: Codable {
 
-    let uri: String
+        let uri: String
        let label: String
        let image: String
        let source: String
        let url: String
        let shareAs: String
-       let yield: Int
+       let yield: Double
        let dietLabels: [String]
        let healthLabels: [String]
        let ingredientLines: [String]
